Use canonical profile route for solvers view
