+++ src/routes/[handle=at_handle]/[shareId]/+page.ts
@@ -2,9 +2,10 @@ import type { PageLoad } from './$types';
 
 export const prerender = false;
 
-export const load: PageLoad = ({ params }) => {
+export const load: PageLoad = ({ params, url }) => {
   return {
     initialActorHandle: params.handle,
-    initialOrderId: params.shareId
+    initialOrderId: params.shareId,
+    taskText: url.searchParams.get('task') ?? ''
   };
 };
